added a function to remove a footstep from the footstep list

+++ CommonWalkingControlModules/src/us/ihmc/commonWalkingControlModules/desiredFootStep/WalkingMessageHandler.java
@@ -288,7 +288,7 @@ public class WalkingMessageHandler
       return transferTime.getDoubleValue() + swingTime.getDoubleValue();
    }
 
-   public double getCurrentNumberOfFootsteps()
+   public int getCurrentNumberOfFootsteps()
    {
       return currentNumberOfFootsteps.getIntegerValue();
    }
